[Bronze I] Title: A/B - 2, Time: 68 ms, Memory: 38512 KB, Score: 38 point -BaekjoonHub

+++ 백준/Bronze/15792. A／B － 2/A／B － 2.py
@@ -1,2 +1,3 @@
-a, b = map(int, input().split())
+from decimal import Decimal
+a, b = map(Decimal, input().split())
 print(a / b)
